fix(analytics): remove redundant float() calls on pandas mean and std division

+++ biogears_service/api/analytics.py
@@ -626,8 +626,8 @@ def compute_time_in_range(csv_path: Path, has_diabetes: bool = False) -> Dict[st
     tar   = (series > hi).sum() / n * 100
     tbr   = (series < lo).sum() / n * 100
     urg   = (series < urgency_lo).sum() / n * 100
-    mean_g = round(float(series.mean()), 1)
-    cv_pct = round(float(series.std() / series.mean() * 100) if series.mean() > 0 else 0, 1)
+    mean_g = round(series.mean(), 1)
+    cv_pct = round(series.std() / series.mean() * 100 if series.mean() > 0 else 0, 1)
 
     return {
         "time_in_range_pct":    round(tir, 1),
